rename item detail list component file
