+++ src/renderer/components/item-list/item-detail-list/item-detail-list.module.css
